// frontend/lib/api/generated/model/transferenciaAlmacenDetalleOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface TransferenciaAlmacenDetalleOut {
  transferencia_detalle_id: number;
  producto_id: number;
  producto_codigo: string;
  producto_nombre: string;
  cantidad_enviada: number;
  cantidad_recibida: number | null;
  diferencia: number;
  observacion_diferencia: string | null;
}